Split utils into separate file
